(refactor) Package separation; Delete view dependancy of domain
