+++ src/main/java/racingcar/domain/RacingRule.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.List;
