+++ src/test/java/racingcar/CollectionsTest.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
 
 public class CollectionsTest {
