+++ src/test/java/racingcar/domain/RaceResultsTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -7,6 +7,9 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.LapResult;
+import racingcar.domain.RaceResults;
 
 public class RaceResultsTest {
